refactor: move pages into view package
